backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bidict==0.23.1
certifi==2025.8.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
prisma==0.15.0
pydantic==1.10.22
pydantic_core==2.33.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.47.2
StrEnum==0.4.15
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wsproto==1.2.0
